# [extension/encoding/awslogs_encoding] Improve performance