… extension/encoding/awslogsencodingextension/internal/unmarshaler/vpc-flow-log/testdata/vpc_flow_log_too_few_fields.log