@constanca-m 2025-04-01
@axw I have changed the naming convention to use `network.local.address` instead of peer, but did not go further than that. In regard to this issue: https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/39051. I changed it here since it was convenient as I was changing the `ips` map to a struct.